+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.2.0](IN PROGRESS)
 
 * Update eslint to v6.2.1 (UIEH-818)
+* Fix view of Action button on package view page header (UIEH-829)
 
 ## [2.1.0](https://github.com/folio-org/ui-eholdings/tree/v2.1.0) (2019-12-04)